requirements.txt
#desarrollo
#reflex==0.8.27

#produccion
reflex==0.8.27
sqlmodel
psycopg2-binary
python-dotenv
bcrypt
alembic 